Inventario Listo, Solo un pequeño detalle Facturas funcionando al 50%

+++ Hawac/src/cr/ac/una/fucem/inge/hawac/model/LineaTableModel.java
@@ -48,7 +48,7 @@ public class LineaTableModel extends AbstractTableModel{
         }
     }
     
-    public float calculaPrecio(Linea l1){
+    public double calculaPrecio(Linea l1){
         float precio = l1.getCantidad()*(float)l1.getProducto().getPrecio();
         return precio;
     }
